Save recovered files and updates on master

+++ src/main/java/com/techn/journalapp/service/JournalEntryService.java
@@ -4,6 +4,8 @@ import com.techn.journalapp.entity.JournalEntry;
 import com.techn.journalapp.entity.User;
 import com.techn.journalapp.repository.JournalEntryRepository;
 import org.bson.types.ObjectId;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -12,8 +14,6 @@ import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
-import java.util.function.Function;
-import java.util.function.Predicate;
 
 @Service
 public class JournalEntryService {
@@ -24,6 +24,7 @@ public class JournalEntryService {
     @Autowired
     private UserService userService;
 
+
     @Transactional
     public void save(JournalEntry journalEntry, String userName){
         try {
@@ -31,10 +32,10 @@ public class JournalEntryService {
             journalEntry.setDate(LocalDateTime.now());
             JournalEntry saved = journalEntryRepository.save(journalEntry);
             userInDb.getJournalEntries().add(saved);
-            userService.saveEntries(userInDb);
+            userService.saveInUser(userInDb);
         } catch (Exception e) {
             System.out.println(e);
-            throw new RuntimeException("illeagel Entry "+e);
+            throw new RuntimeException("Illeagel Entry "+e);
         }
     }
 
@@ -50,27 +51,27 @@ public class JournalEntryService {
         return  journalEntryRepository.findById(objectId);
     }
 
-    public void deleteById(ObjectId objectId, String userName){
-        User userInDb = userService.findByUserName(userName);
+    @Transactional
+    public boolean deleteById(ObjectId objectId, String userName){
+        boolean removed= false;
+        try {
+            User userInDb = userService.findByUserName(userName);
 
-        // 1. Delete the journal entry from DB
+            // 1 . Remove the journal entry from User's List<JournalEntry>
+            removed = userInDb.getJournalEntries().removeIf((x) -> x.getId().equals(objectId));
 
+            if(removed) {
+                // 2. Delete the journal entry from DB
+                journalEntryRepository.deleteById(objectId);
 
-        // 2. Remove the journal entry from user's list manually
-//        List<JournalEntry> updatedList = new ArrayList<>();
-//        for (JournalEntry entry : userInDb.getJournalEntries()) {
-//            if (!entry.getId().equals(objectId)) {
-//                updatedList.add(entry);
-//            }
-//        }
-
-//        userInDb.setJournalEntries(updatedList);
-        userInDb.getJournalEntries().removeIf((x)-> x.getId().equals(objectId));
-
-        journalEntryRepository.deleteById(objectId);
-
-        // 3. Save updated user
-        userService.saveEntries(userInDb);
+                // 3. Save updated user
+                userService.saveInUser(userInDb);
+            }
+        } catch (Exception e) {
+            System.out.println(e);
+            throw new RuntimeException("Something went wrong"+e);
+        }
+        return removed;
     }
 
 }
